Fixed issue #93

+++ Classes/JBBarChartView.h
@@ -111,7 +111,7 @@
  *
  *  @return Horizontal width (in pixels) between each bar.
  */
-- (NSUInteger)barPaddingForBarChartView:(JBBarChartView *)barChartView;
+- (CGFloat)barPaddingForBarChartView:(JBBarChartView *)barChartView;
 
 @end
